requirements.txt
amqp==2.1.4
asn1crypto==0.22.0
beautifulsoup4==4.6.0
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
cryptography==1.9
Django==1.11.2
django-bootstrap-form==3.2.1
django-celery-beat==1.0.1
django-markdown-deux==1.0.5
django-model-utils==3.0.0
django-notifications-hq==1.2
django-simple-captcha==0.5.5
django-sslserver==0.19
djangorestframework==3.6.3
enum34==1.1.6
future==0.16.0
idna==2.5
ipaddress==1.0.18
jsonfield==2.0.1
kombu==4.0.2
mailjet==1.4.0
mailjet-rest==1.3.0
markdown2==2.3.4
MySQL-python==1.2.5
nexmo==1.5.0
olefile==0.44
Pillow==4.1.1
poster==0.8.1
pycparser==2.17
PyJWT==1.5.2
python-telegram-bot==6.1.0
pytz==2017.2
redis==2.10.5
requests==2.14.2
six==1.10.0
Unirest==1.1.7
vine==1.1.3
